requirements.txt
streamlit>=1.25.0
nltk>=3.8
scikit-learn>=1.2
matplotlib>=3.7
wordcloud>=1.9
pdfplumber>=0.10
python-docx>=0.8.11
